Add exhaustive NAD83 -> NAD83(HPGN) shift grids
The main proj-datumgrid package contains 5 grids (FL, MD, TN, WI, WO)
out of the full 51 grids for the 51 states. So we provide them in an
extensive way by using the build_nad83_harn.sh script that downloads
the source .los/.las files, and use the scripts/loslas2ntv2.py script
to convert them to NTv2.

Important note: https://download.osgeo.org/proj/hpgn_ntv2.zip contains
a similar content, except that this historical package had the sign of
the offsets reverse (so doing NAD83(HPGN) -> NAD83 conversion), because
of the requirements of the historical +nadgrids, that required grids to
go from a source CRS to WGS84 (here NAD83 was considered as being WGS84)
With the new +proj=hgridshift operator, we don't need such hacks, and
it is better to keep the grids as close as possible as their sources.