Regenerate g2018u0.gtx and g2018p0.gtx
The previous versions were generated on November 23 2019, but
as pointed out by @elil in
https://github.com/OSGeo/proj-datumgrid/issues/55#issuecomment-564148596 ,
NOAA has issued a warning notice:

```
Urgent: Critical Change to GEOID18 grid files**

NGS has corrected errors in some of the GEOID18 grid files posted on the
GEOID18 Downloads webpage prior to November 26, 2019. See below for more
information and details:

Critical Change

The GEOID18 grid files in the little endian and ASCII grid formats that were
posted on the NGS website prior to November 26, 2019, contained errors in a
small percentage of the grid cells (1 in 500). NGS has since corrected the
errors and replaced these files on the GEOID18 download page. We encourage
users that have previously downloaded those grids to immediately replace them
with the updated grids.
```

We do use the little endian files as the source, so we were affected.