Fixing horizontal resolution of New Zealand .gtx grids
The New Zealand grids .gtx grids were converted from source files for which the numerical precision of the grid spacing was insufficient (about 6dp).  This is fixed in these files to ensure the correct location of points.

As an example to spacing of nzgeoid2016.gtx is 0.0166667 degrees which
multiplied by the number of grid rows/columns could give an error in
location of over 7m. The corresponding error in geoid height will be
less than 1mm.